enhance: Add Describe User methods (#697)
resolves: #698

I've added an entity:

```
type UserDescription struct {
	Name  string
	Roles []string
}
```

And two methods and tests for them:

- `DescribeUser(ctx context.Context, username string)
(entity.UserDescription, error)` - which returns UserDescription for a
specific user.
- `DescribeUsers(ctx context.Context) ([]entity.UserDescription, error)`
- which returns UserDescription for all users presented.

The thing that I want to raise is that I am not sure if it is feasible
to return empty `UserDescription` entity in case the user from
`DescribeUser` method doesn't exist, I would highly appreciate
suggestions here.

Also, I am not much familiar with the `milvus-sdk-go` codebase and this
is my first PR here, so any comments/feedback/suggestions would be
highly appreciated, thanks!

---------

Signed-off-by: punkerpunker <gleb.vazhenin@team.bumble.dev>
Co-authored-by: punkerpunker <gleb.vazhenin@team.bumble.dev>