fix broken Grant() call (#784)
The latest version of the SDK doesn't pass the actual privilege that is
granted during a Grant() call. This causes the Grant call to fail in any
situation.

This PR fixes this issue by extending the method signature to take a
`privilege` parameter, similarly to how the Python SDK does. This was an
easy fix because the protobuf structs already had the field in place, so
the SDK only needs to use that correct field in the API call.

Signed-off-by: Alex Giurgiu <alex@giurgiu.io>